[ROUTES] Arduino routes separated from User's ones
